Restructure directories in preparation for PyPI release
Also:
- add devcheck script to run the sanity checks: tests and examples
- order imports correctly
