+++ examples/convention-tag/language.test.yaml
@@ -16,19 +16,23 @@ test:
   suites:
   - name:  "Language samples test"
     setup:
-      - log:
-        - 'In setup {}'
-        - hi
+      - code:
+          log('In setup')
     teardown:
       - code:
-          log('In teardown {}', 'bye')
+          log('In teardown')
     cases:
     - name: Sentiment from text, direct call
       spec:
       - code: |
-          out = call("python3 examples/testdata/googleapis/artman-genfiles/python/language-v1/samples/google/cloud/language_v1/gapic/analyze_sentiment/analyze_sentiment_request_language_sentiment_text.py", content="happy happy smile hope")
-          assert_success()
+          out = call("python3 examples/mock_samples/googleapis/artman-genfiles/python/language-v1/samples/google/cloud/language_v1/gapic/analyze_sentiment/analyze_sentiment_request_language_sentiment_text.py", content="happy happy smile hope")
 
+      # the following shows how to interleave yaml and code
+      - assert_success:
+        - "that should have worked {}"  # try `assert_failure` to see message
+        - well
+
+      - code: |
           import re
 
           # TODO: still need a way to do text matching in a declarative test like we do below, inside `code`.
@@ -37,17 +41,19 @@ test:
           assert_that(score_found is not None, 'score matches regexp')
           score = float(score_found.group(1))
           assert_that(score > 0.7, 'score is high')
+          home = env('HOME')
+          log('home directory: {}'.format(home))
 
           magnitude_found = re.search('magnitude: ([0123456789.]+)', out)
           assert_that(magnitude_found is not None, 'magnitude matches regexp')
           magnitude = float(magnitude_found.group(1))
           assert_that(magnitude > 0.7, 'magnitude is high')
 
-    - name: Sentiment from text, API call convention
+    - name: Sentiment from text, manifest call convention
       spec:
       - code: |
-          out = call("Google.Cloud.Language.v1.AnalyzeSentiment:analyze_sentiment_request_language_sentiment_text", content="happy happy smile hope")
-
+          out = call("language_analyze_sentiment_text", content="happy happy smile hope")
+          assert_success("that should have worked", "well")
           import re
 
           # TODO: still need a way to do text matching in a declarative test like we do below, inside `code`.
@@ -61,3 +67,23 @@ test:
           assert_that(magnitude_found is not None, 'magnitude matches regexp')
           magnitude = float(magnitude_found.group(1))
           assert_that(magnitude > 0.7, 'magnitude is high')
+
+          assert_not_contains("random message", "the rain in Spain")
+          
+    - name: Sentiment from text, manifest call convention, yaml directives
+      spec:
+      - call:
+          target: "language_analyze_sentiment_text"
+          params:
+            content:
+              literal: "happy happy smile hope"
+      - assert_success: [] # try assert_failure
+      - assert_contains:
+          - message: "Score is very positive"
+          - literal: "score: 0.8"
+      - assert_contains:
+          - message: "Magnitude is very positive"
+          - literal: "magnitude: 0.8"
+      - assert_not_contains:
+          - message: "Random message"
+          - literal: "The rain in Spain falls mainly in the plain"
