+++ examples/mock_samples/googleapis/artman-genfiles/go/language-v1/samples/google/cloud/language_v1/gapic/analyze_sentiment/analyze_sentiment.go
