+++ examples/mock_samples/samples/vision/product-search/create_product.py
