+++ examples/mock_samples/samples/vision/product-search/delete_product.py
