+++ examples/mock_samples/samples/vision/product-search/list_products.py
