+++ examples/sample.manifest.yaml
