+++ src/convention/__init__.py
@@ -25,8 +25,9 @@ all_files = [entry for entry in os.listdir(__abs_file_path__)
              if entry !='__init__.py' and entry != '__pycache__' and not entry.endswith('~')]
 all_conventions = [os.path.splitext(os.path.basename(entry))[0] for entry in all_files]
 environment_creators = {}
+print("All conventions: {}".format(all_conventions))
 for convention in all_conventions:
-  module = importlib.import_module('.'+convention, package='convention')
+  module = importlib.import_module('.'+convention, package='src.convention')
   if 'test_environments' in dir(module):
     logging.info('registering convention "{}"'.format(convention))
     environment_creators[convention] = module.test_environments
