+++ src/convention/cloud/cloud.py
@@ -12,11 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import testenv
-
-import os
 import re
+import os
 
+from src import testenv
 
 class PythonSamplesEnvironment(testenv.Base):
